feat: distribute via slots
Moving the assigned nodes from the slot directly to the bottom-bar is not compatible with lit-html.
When lit-html tries to remove the elements it rendered, it will fail to remove the moved DOM, thus
leading to duplicate items in the bottom-bar.

BREAKING CHANGE: If your component allows adding content to the bottom bar via slots, then you have
to also make sure you also export `bottom-bar-toolbar` and `bottom-bar-menu`, otherwise the actions
will not show up.

Fixes #24802 (RM:24802)